FIXED: Correct remaining uses of 32-bit indexes.
The goTo() and skip() functions take `size_t' i.o. `unsigned int'
arguments in order to support 64-bit indexes.

+++ libhdt/src/hdt/TripleIDStringIterator.cpp
@@ -50,7 +50,7 @@ size_t TripleIDStringIterator::estimatedNumResults() {
 ResultEstimationType TripleIDStringIterator::numResultEstimation() {
 	return iterator->numResultEstimation();
 }
-void TripleIDStringIterator::skip(unsigned int pos) {
+void TripleIDStringIterator::skip(size_t pos) {
 	iterator->skip(pos);
 }
